Implemented new state diagram
